backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]          
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
email-validator              
jinja2